Cleaned up propagators after crash.
