+++ propulate/propagators/__init__.py
@@ -4,10 +4,6 @@ This package holds all Propagator subclasses including the Propagator itself.
 
 __all__ = ["Propagator", "Stochastic", "Conditional", "Compose", "PointMutation", "RandomPointMutation",
            "IntervalMutationNormal", "MateUniform", "MateMultiple", "MateSigmoid", "SelectMin", "SelectMax",
-           "SelectUniform", "InitUniform", "PSOPropagator", "PSOInitUniform"]
+           "SelectUniform", "InitUniform"]
 
 from propulate.propagators.propagators import *
-from propulate.propagators.pso_propagator import PSOPropagator
-from propulate.propagators.init_propagators import InitUniform, PSOInitUniform
-
-
